+++ lizard_ui/tests.py
@@ -4,7 +4,6 @@ from django.contrib.auth.models import User
 from django.core.urlresolvers import reverse
 from django.test import TestCase
 from django.test.client import RequestFactory
-from django.utils import simplejson as json
 
 from lizard_ui.models import ApplicationIcon
 from lizard_ui.configchecker import checker